refactor: move cache to eodag.utils
